setpoints scale issue fix
The yaw and pitch states can be nonzero at trim, depending on their setpoints, so we have to account that in. Especialy yaw was giving very small inputs cos heading is any value that means the logic wasnt valid. This means I had to move this function after the Del_y calculation in the control_fun, and also renamed it cos we scale only the omg setpoints here.